fix(engine): PR #796 review - moving global flag into config.ts
