+++ packages/lwc-engine/src/framework/config.ts
@@ -1,0 +1,9 @@
+export let isSyntheticShadowRoot: boolean;
+
+export function alwaysUseSyntheticShadowRoot() {
+    isSyntheticShadowRoot = true;
+}
+
+export function alwaysUseNativeShadowRoot() {
+    isSyntheticShadowRoot = false;
+}
