+++ packages/lwc-engine/src/framework/patch.ts
@@ -6,7 +6,7 @@ import { setPrototypeOf, create, isUndefined } from "../shared/language";
 import { ComponentDef } from "./def";
 import { HTMLElementConstructor } from "./base-bridge-element";
 import { PatchedElement, PatchedSlotElement, PatchedNode, PatchedIframeElement, PatchedCustomElement } from '../faux-shadow/faux';
-import { isSyntheticShadowRoot } from "./vm";
+import { isSyntheticShadowRoot } from "./config";
 
 // Using a WeakMap instead of a WeakSet because this one works in IE11 :(
 const FromIteration: WeakMap<VNodes, 1> = new WeakMap();
